requirements.txt
Django==2.0.2
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.3
mysqlclient==1.3.12
pygments==2.2.0
django-guardian==1.4.9
pillow==5.3.0